Sonic can now jump from lower ground onto higher ground

+++ src/uk/co/jamesj999/sonic/sprites/managers/PlayableSpriteMovementManager.java
@@ -94,7 +94,7 @@ public class PlayableSpriteMovementManager extends
 			sprite.setAir(true);
 			ySpeed += sprite.getJump();
 		}
-		byte height = terrainCollisionManager.calculateTerrainHeight(sprite);
+		short height = terrainCollisionManager.calculateTerrainHeight(sprite);
 
 		if (sprite.getAir()) {
 			ySpeed -= sprite.getGravity();
